better use internal HSI oscillator     rcc_clock_setup_pll(&rcc_hsi_configs[RCC_CLOCK_3V3_84MHZ]);     systick_set_frequency(84000, 84000);     systick_set_reload(84000);     systick_counter_enable();     systick_interrupt_enable();
-> https://os.mbed.com/questions/70461/Why-does-this-nucleo-board-lack-a-crysta/
Mine has MB1136 C-01
see (page 24/69 UM1724 Rev 13, en.DM00105823.pdf)

"There are two possible default configurations of the HSE pins, depending on the version of
the STM32 Nucleo board hardware.
The board version MB1136 C-01 or MB1136 C-02 is mentioned on the sticker, placed on the
bottom side of the PCB.
The board marking MB1136 C-01 corresponds to a board, configured as HSE not used.
The board marking MB1136 C-02 (or higher) corresponds to a board, configured to use
ST-LINK MCO as clock input."